Get smoothing working!
Thanks to forerunrun: https://discourse.threejs.org/t/smooth-shading-on-buffergeometry/27582/2

It still has artifacts in places that should be hard edges (or nearly hard edges), like on the bishop and rook, but there's no more gaps!

A better solution would be to edit the 3D models in an external program to compute good normals, or edit the C code to generate models with normals, or find new 3D models.