Fix cube wrapping partially
Pieces can no longer move off into the air, at least!
But the orientation of movement doesn't translate to the other side properly, so a rook can veer off to the side once wrapping around.
Oh yeah, I was gonna limit the movement to going one over the edge anyways, so that should... fix it but not for the knight, right?