Fix case where it's possible to Assassin-Mate
When checking whether a move would put a player in check, for moves that
re-orient a piece onto a wall without moving it to another space, it was
supposing that the piece didn't exist at the location it was originally
occupying, in priority over supposing that there's a piece at the
location at the end of the move, and thus, was supposing that the piece
didn't exist in the hypothetical board. So I've swapped the priorities.