WIP: convert game-audio.js to ESM
Firefox is giving me: "Uncaught SyntaxError: export declarations may only appear at top level of a module"
which doesn't make sense. It is at top level.
Chrome says: "Uncaught SyntaxError: Unexpected token 'export'"... as well as "The AudioContext was not allowed to start. It must be resumed (or created) after a user gesture on the page."
Both on line 4 of game-audio.js; if there's a syntax error, why is the code executing? Riddle me that.